+++ packages/classes/src/SDKService.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2020 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License version 3
@@ -15,17 +15,27 @@
  */
 
 import { Observable } from 'rxjs';
+import { fromFetch } from 'rxjs/fetch';
 import { pluck } from 'rxjs/operators';
-import { ajax, AjaxResponse } from 'rxjs/ajax';
-import 'url-search-params-polyfill';
+import { ajax } from 'rxjs/ajax';
+import { LookupTable } from '@craftercms/models';
+import { stringify } from 'query-string';
+import { crafterConf } from './config';
 
-export function httpGet<T extends any = any>(requestURL: string, params: Object = {}): Observable<T> {
-  const searchParams = new URLSearchParams(params as URLSearchParams);
-  return ajax.get(`${requestURL}?${searchParams.toString()}`).pipe(pluck<AjaxResponse, T>('response'));
+export function httpGet<T extends any = any>(requestURL: string, params: Record<string, any> = {}, headers?: LookupTable): Observable<T> {
+  const mode = crafterConf.getConfig().cors;
+  return fromFetch(`${requestURL}?${stringify(params)}`, {
+    method: 'GET',
+    headers: headers,
+    mode: typeof mode === 'boolean' ? mode ? 'cors' : 'no-cors' : mode,
+    selector: response => response.json()
+  });
 }
 
-export function httpPost<T extends any = any>(requestURL: string, body: Object = {}): Observable<T> {
-  return ajax.post(requestURL, body, { 'Content-Type': 'application/json' }).pipe(pluck<AjaxResponse, T>('response'));
+export function httpPost<T extends any = any>(requestURL: string, body: Object = {}, headers?: LookupTable): Observable<T> {
+  return ajax.post(requestURL, body, { 'Content-Type': 'application/json', ...headers }).pipe(
+    pluck('response')
+  );
 }
 
 export const SDKService = {