+++ packages/content/src/UrlTransformationService.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2020 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License version 3
@@ -14,27 +14,39 @@
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
 
-import { Observable } from 'rxjs';
-
 import { crafterConf, SDKService } from '@craftercms/classes';
 import { CrafterConfig } from '@craftercms/models';
 import { composeUrl } from '@craftercms/utils';
+import { Observable } from 'rxjs';
+
+export type UrlTransformers =
+  | 'storeUrlToRenderUrl'
+  | 'renderUrlToStoreUrl'
+  | 'renderUrlToTargetedStoreUrl'
+  | 'storeUrlToFullRenderUrl'
+  | 'toWebAppRelativeUrl'
+  | 'toServletRelativeUrl'
+  | 'toFullUrl'
+  | 'toFullHttpsUrl'
+  | 'folderToIndexUrl'
+  | 'toTargetedUrl'
+  | 'toCurrentTargetedUrl';
 
 /**
  * Transforms a URL, based on the current site's UrlTransformationEngine.
  * @param {string} transformerName - Name of the transformer to apply
  * @param {string} url - URL that will be transformed
  */
-export function urlTransform(transformerName: string, url: string): Observable<any>;
-export function urlTransform(transformerName: string, url: string, config: CrafterConfig): Observable<any>;
-export function urlTransform(transformerName: string, url: string, config?: CrafterConfig): Observable<any> {
+export function urlTransform(transformerName: UrlTransformers, url: string): Observable<string>;
+export function urlTransform(transformerName: UrlTransformers, url: string, config: CrafterConfig): Observable<string>;
+export function urlTransform(transformerName: UrlTransformers, url: string, config?: CrafterConfig): Observable<string> {
   config = crafterConf.mix(config);
   const requestURL = composeUrl(config, config.endpoints.TRANSFORM_URL);
-  return SDKService.httpGet(requestURL, {
+  return SDKService.httpGet<string>(requestURL, {
     crafterSite: config.site,
     transformerName,
     url
-  });
+  }, config.headers);
 }
 
 /**