+++ packages/models/src/NavigationItem.ts
@@ -14,4 +14,13 @@
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
 
-export * from './hooks';
+export interface NavigationItem {
+  url: string;
+  label: string;
+  active: boolean;
+  subItems: Array<NavigationItem>;
+  attributes: Record<string, string>;
+  [prop: string]: any;
+}
+
+export default NavigationItem;